Rename C++ converter test
