--- tools/a2mochi/x/cpp/convert_test.go
@@ -1,141 +1,0 @@
-//go:build slow
-
-package cpp_test
-
-import (
-	"bytes"
-	"flag"
-	"os"
-	"os/exec"
-	"path/filepath"
-	"strings"
-	"testing"
-
-	"mochi/ast"
-	"mochi/parser"
-	"mochi/runtime/vm"
-	"mochi/types"
-
-	"mochi/tools/a2mochi/x/cpp"
-)
-
-var update = flag.Bool("update", false, "update golden files")
-
-func findRepoRoot(t *testing.T) string {
-	dir, err := os.Getwd()
-	if err != nil {
-		t.Fatal(err)
-	}
-	for i := 0; i < 10; i++ {
-		if _, err := os.Stat(filepath.Join(dir, "go.mod")); err == nil {
-			return dir
-		}
-		parent := filepath.Dir(dir)
-		if parent == dir {
-			break
-		}
-		dir = parent
-	}
-	t.Fatal("go.mod not found")
-	return ""
-}
-
-func runMochi(src string) ([]byte, error) {
-	prog, err := parser.ParseString(src)
-	if err != nil {
-		return nil, err
-	}
-	env := types.NewEnv(nil)
-	if errs := types.Check(prog, env); len(errs) > 0 {
-		return nil, errs[0]
-	}
-	p, err := vm.Compile(prog, env)
-	if err != nil {
-		return nil, err
-	}
-	var out bytes.Buffer
-	m := vm.New(p, &out)
-	if err := m.Run(); err != nil {
-		return nil, err
-	}
-	return bytes.TrimSpace(out.Bytes()), nil
-}
-
-func TestConvert_Golden(t *testing.T) {
-	if _, err := exec.LookPath("clang++"); err != nil {
-		t.Skipf("clang++ not installed: %v", err)
-	}
-
-	root := findRepoRoot(t)
-	pattern := filepath.Join(root, "tests/transpiler/x/cpp", "*.cpp")
-	files, err := filepath.Glob(pattern)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if len(files) == 0 {
-		t.Fatalf("no files: %s", pattern)
-	}
-
-	outDir := filepath.Join(root, "tests/a2mochi/x/cpp")
-	os.MkdirAll(outDir, 0o755)
-
-	for _, srcPath := range files {
-		name := strings.TrimSuffix(filepath.Base(srcPath), ".cpp")
-		t.Run(name, func(t *testing.T) {
-			data, err := os.ReadFile(srcPath)
-			if err != nil {
-				t.Fatalf("read src: %v", err)
-			}
-			node, err := cpp.Parse(string(data))
-			if err != nil {
-				t.Fatalf("parse: %v", err)
-			}
-			_, err = cpp.ConvertSource(node)
-			if err != nil {
-				t.Fatalf("convert source: %v", err)
-			}
-			astNode, err := cpp.Convert(string(data))
-			if err != nil {
-				t.Fatalf("convert: %v", err)
-			}
-			got := []byte(astNode.String())
-			outPath := filepath.Join(outDir, name+".ast")
-			if *update {
-				os.WriteFile(outPath, got, 0644)
-			}
-			want, err := os.ReadFile(outPath)
-			if err != nil {
-				t.Fatalf("missing golden: %v", err)
-			}
-			if string(got) != string(want) {
-				t.Fatalf("golden mismatch\n--- Got ---\n%s\n--- Want ---\n%s", got, want)
-			}
-
-			var buf bytes.Buffer
-			if err := ast.Fprint(&buf, astNode); err != nil {
-				t.Fatalf("print: %v", err)
-			}
-			code := buf.String()
-			mochiPath := filepath.Join(outDir, name+".mochi")
-			if *update {
-				os.WriteFile(mochiPath, []byte(code), 0644)
-			}
-
-			gotOut, err := runMochi(code)
-			if err != nil {
-				t.Fatalf("run: %v", err)
-			}
-			vmSrc, err := os.ReadFile(filepath.Join(root, "tests", "vm", "valid", name+".mochi"))
-			if err != nil {
-				t.Fatalf("missing vm source: %v", err)
-			}
-			wantOut, err := runMochi(string(vmSrc))
-			if err != nil {
-				t.Fatalf("run vm: %v", err)
-			}
-			if !bytes.Equal(gotOut, wantOut) {
-				t.Fatalf("output mismatch\nGot: %s\nWant: %s", gotOut, wantOut)
-			}
-		})
-	}
-}
