+++ tools/a2mochi/x/cpp/print.go
@@ -1,0 +1,47 @@
+package cpp
+
+import (
+	"fmt"
+	"os"
+	"path/filepath"
+	"runtime"
+	"strings"
+	"time"
+
+	"mochi/ast"
+)
+
+var version = func() string {
+	_, file, _, ok := runtime.Caller(0)
+	if ok {
+		data, err := os.ReadFile(filepath.Join(filepath.Dir(file), "../../../../VERSION"))
+		if err == nil {
+			return strings.TrimSpace(string(data))
+		}
+	}
+	return "unknown"
+}()
+
+func header() string {
+	tz := time.FixedZone("GMT+7", 7*60*60)
+	return fmt.Sprintf("// Generated by a2mochi v%s on %s", strings.TrimSpace(version), time.Now().In(tz).Format("2006-01-02 15:04:05 MST"))
+}
+
+// Print converts an AST node into Mochi source code.
+func Print(n *ast.Node) (string, error) {
+	if n == nil {
+		return "", fmt.Errorf("nil node")
+	}
+	var code strings.Builder
+	if err := ast.Fprint(&code, n); err != nil {
+		return "", err
+	}
+	var b strings.Builder
+	b.WriteString(header())
+	b.WriteByte('\n')
+	b.WriteString(code.String())
+	if !strings.HasSuffix(code.String(), "\n") {
+		b.WriteByte('\n')
+	}
+	return b.String(), nil
+}
